refactor(bgrec): split Testing Support region into BackgroundRecorder.Testing.cs
